+++ src/common/ast/mod.rs
@@ -1,2 +1,2 @@
-pub mod ast;
+mod ast;
 pub use ast::*;
