+++ src/lib.rs
@@ -1,0 +1,2 @@
+mod common;
+pub use common::ast;
